+++ src/models/Section.php
@@ -34,6 +34,8 @@ class Section extends BaseSection
     public const EVENT_DEFINE_IMAGE = 'defineImageEvent';
     public const EVENT_DEFINE_ICON = 'defineIconEvent';
 
+    // ATTENTION! If types are changed here, be sure to update the section scaffold template!
+
     public array $actions = [];
     public bool $allSites = false;
     public array|string $entryType = '';
@@ -862,6 +864,11 @@ class Section extends BaseSection
                     ->provisionalDrafts(true);
                 break;
             }
+            case 'ownProvisional':
+                $query
+                    ->provisionalDrafts(true)
+                    ->draftCreator(Craft::$app->user->identity);
+                break;
             case 'all':
             {
                 $query